os: updated return types to match 'syscall'

+++ os/unix/file_unix.go
@@ -10,10 +10,10 @@ func Close(f int32) error {
 	return syscall.Close(f)
 }
 
-func Read(f int32, buf []byte) (int64, error) {
+func Read(f int32, buf []byte) (int, error) {
 	return syscall.Read(f, buf)
 }
 
-func Write(f int32, buf []byte) (int64, error) {
+func Write(f int32, buf []byte) (int, error) {
 	return syscall.Write(f, buf)
 }
